algo_trading_strat/monitoring_screen/requirements.txt
alpaca-py==0.42.0
numpy==2.2.6
pandas==2.3.0
plotly==6.2.0
streamlit==1.47.1
